Move github to internal/collector/
