+++ internal/collector/github/factory.go
