+++ internal/collector/github/legacy/constants.go
