+++ internal/collector/github/legacy/issues.go
