+++ internal/collector/github/legacy/util.go
