+++ internal/collector/github/queries.go
